rename date-picker class names
