+++ vaadin-date-picker-flow-parent/vaadin-date-picker-flow-integration-tests/src/test/java/com/vaadin/flow/component/datepicker/validation/BasicValidationIT.java
@@ -7,33 +7,52 @@ import com.vaadin.tests.validation.AbstractValidationIT;
 import org.junit.Test;
 import org.openqa.selenium.Keys;
 
-import static com.vaadin.flow.component.datepicker.validation.DatePickerValidationBinderPage.MIN_INPUT;
-import static com.vaadin.flow.component.datepicker.validation.DatePickerValidationBinderPage.MAX_INPUT;
-import static com.vaadin.flow.component.datepicker.validation.DatePickerValidationBinderPage.EXPECTED_VALUE_INPUT;
-import static com.vaadin.flow.component.datepicker.validation.DatePickerValidationBinderPage.REQUIRED_ERROR_MESSAGE;
-import static com.vaadin.flow.component.datepicker.validation.DatePickerValidationBinderPage.UNEXPECTED_VALUE_ERROR_MESSAGE;
+import static com.vaadin.flow.component.datepicker.validation.BasicValidationPage.MIN_INPUT;
+import static com.vaadin.flow.component.datepicker.validation.BasicValidationPage.MAX_INPUT;
+import static com.vaadin.flow.component.datepicker.validation.BasicValidationPage.REQUIRED_BUTTON;
 
-@TestPath("vaadin-date-picker/validation/binder")
-public class DatePickerValidationBinderIT
+@TestPath("vaadin-date-picker/validation/basic")
+public class BasicValidationIT
         extends AbstractValidationIT<DatePickerElement> {
     @Test
     public void fieldIsInitiallyValid() {
         assertClientValid();
         assertServerValid();
-        assertErrorMessage(null);
+    }
+
+    @Test
+    public void detach_attach_preservesInvalidState() {
+        // Make field invalid
+        $("button").id(REQUIRED_BUTTON).click();
+        testField.sendKeys(Keys.TAB);
+
+        detachAndReattachField();
+
+        assertServerInvalid();
+        assertClientInvalid();
+    }
+
+    @Test
+    public void webComponentCanNotModifyInvalidState() {
+        assertWebComponentCanNotModifyInvalidState();
+
+        detachAndReattachField();
+
+        assertWebComponentCanNotModifyInvalidState();
     }
 
     @Test
     public void required_triggerInputBlur_assertValidity() {
+        $("button").id(REQUIRED_BUTTON).click();
+
         testField.sendKeys(Keys.TAB);
         assertServerInvalid();
         assertClientInvalid();
-        assertErrorMessage(REQUIRED_ERROR_MESSAGE);
     }
 
     @Test
     public void required_changeInputValue_assertValidity() {
-        $("input").id(EXPECTED_VALUE_INPUT).sendKeys("2022-01-01", Keys.ENTER);
+        $("button").id(REQUIRED_BUTTON).click();
 
         testField.setInputValue("1/1/2022");
         assertServerValid();
@@ -42,27 +61,20 @@ public class DatePickerValidationBinderIT
         testField.setInputValue("");
         assertServerInvalid();
         assertClientInvalid();
-        assertErrorMessage(REQUIRED_ERROR_MESSAGE);
     }
 
     @Test
     public void min_changeInputValue_assertValidity() {
         $("input").id(MIN_INPUT).sendKeys("2022-03-01", Keys.ENTER);
-        $("input").id(EXPECTED_VALUE_INPUT).sendKeys("2022-04-01", Keys.ENTER);
 
-        // Constraint validation fails:
         testField.setInputValue("2/1/2022");
         assertClientInvalid();
         assertServerInvalid();
-        assertErrorMessage("");
 
-        // Binder validation fails:
         testField.setInputValue("3/1/2022");
-        assertClientInvalid();
-        assertServerInvalid();
-        assertErrorMessage(UNEXPECTED_VALUE_ERROR_MESSAGE);
+        assertClientValid();
+        assertServerValid();
 
-        // Both validations pass:
         testField.setInputValue("4/1/2022");
         assertClientValid();
         assertServerValid();
@@ -71,21 +83,15 @@ public class DatePickerValidationBinderIT
     @Test
     public void max_changeInputValue_assertValidity() {
         $("input").id(MAX_INPUT).sendKeys("2022-03-01", Keys.ENTER);
-        $("input").id(EXPECTED_VALUE_INPUT).sendKeys("2022-02-01", Keys.ENTER);
 
-        // Constraint validation fails:
         testField.setInputValue("4/1/2022");
         assertClientInvalid();
         assertServerInvalid();
-        assertErrorMessage("");
 
-        // Binder validation fails:
         testField.setInputValue("3/1/2022");
-        assertClientInvalid();
-        assertServerInvalid();
-        assertErrorMessage(UNEXPECTED_VALUE_ERROR_MESSAGE);
+        assertClientValid();
+        assertServerValid();
 
-        // Both validations pass:
         testField.setInputValue("2/1/2022");
         assertClientValid();
         assertServerValid();
@@ -93,12 +99,9 @@ public class DatePickerValidationBinderIT
 
     @Test
     public void badInput_changeInputValue_assertValidity() {
-        $("input").id(EXPECTED_VALUE_INPUT).sendKeys("2022-01-01", Keys.ENTER);
-
         testField.setInputValue("INVALID");
         assertServerInvalid();
         assertClientInvalid();
-        assertErrorMessage("");
 
         testField.setInputValue("1/1/2022");
         assertServerValid();
@@ -107,7 +110,6 @@ public class DatePickerValidationBinderIT
         testField.setInputValue("INVALID");
         assertServerInvalid();
         assertClientInvalid();
-        assertErrorMessage("");
     }
 
     protected DatePickerElement getTestField() {
