+++ vaadin-date-picker-flow-parent/vaadin-date-picker-flow-integration-tests/src/test/java/com/vaadin/flow/component/datepicker/validation/BinderValidationIT.java
@@ -7,52 +7,33 @@ import com.vaadin.tests.validation.AbstractValidationIT;
 import org.junit.Test;
 import org.openqa.selenium.Keys;
 
-import static com.vaadin.flow.component.datepicker.validation.DatePickerValidationBasicPage.MIN_INPUT;
-import static com.vaadin.flow.component.datepicker.validation.DatePickerValidationBasicPage.MAX_INPUT;
-import static com.vaadin.flow.component.datepicker.validation.DatePickerValidationBasicPage.REQUIRED_BUTTON;
+import static com.vaadin.flow.component.datepicker.validation.BinderValidationPage.MIN_INPUT;
+import static com.vaadin.flow.component.datepicker.validation.BinderValidationPage.MAX_INPUT;
+import static com.vaadin.flow.component.datepicker.validation.BinderValidationPage.EXPECTED_VALUE_INPUT;
+import static com.vaadin.flow.component.datepicker.validation.BinderValidationPage.REQUIRED_ERROR_MESSAGE;
+import static com.vaadin.flow.component.datepicker.validation.BinderValidationPage.UNEXPECTED_VALUE_ERROR_MESSAGE;
 
-@TestPath("vaadin-date-picker/validation/basic")
-public class DatePickerValidationBasicIT
+@TestPath("vaadin-date-picker/validation/binder")
+public class BinderValidationIT
         extends AbstractValidationIT<DatePickerElement> {
     @Test
     public void fieldIsInitiallyValid() {
         assertClientValid();
         assertServerValid();
-    }
-
-    @Test
-    public void detach_attach_preservesInvalidState() {
-        // Make field invalid
-        $("button").id(REQUIRED_BUTTON).click();
-        testField.sendKeys(Keys.TAB);
-
-        detachAndReattachField();
-
-        assertServerInvalid();
-        assertClientInvalid();
-    }
-
-    @Test
-    public void webComponentCanNotModifyInvalidState() {
-        assertWebComponentCanNotModifyInvalidState();
-
-        detachAndReattachField();
-
-        assertWebComponentCanNotModifyInvalidState();
+        assertErrorMessage(null);
     }
 
     @Test
     public void required_triggerInputBlur_assertValidity() {
-        $("button").id(REQUIRED_BUTTON).click();
-
         testField.sendKeys(Keys.TAB);
         assertServerInvalid();
         assertClientInvalid();
+        assertErrorMessage(REQUIRED_ERROR_MESSAGE);
     }
 
     @Test
     public void required_changeInputValue_assertValidity() {
-        $("button").id(REQUIRED_BUTTON).click();
+        $("input").id(EXPECTED_VALUE_INPUT).sendKeys("2022-01-01", Keys.ENTER);
 
         testField.setInputValue("1/1/2022");
         assertServerValid();
@@ -61,20 +42,27 @@ public class DatePickerValidationBasicIT
         testField.setInputValue("");
         assertServerInvalid();
         assertClientInvalid();
+        assertErrorMessage(REQUIRED_ERROR_MESSAGE);
     }
 
     @Test
     public void min_changeInputValue_assertValidity() {
         $("input").id(MIN_INPUT).sendKeys("2022-03-01", Keys.ENTER);
+        $("input").id(EXPECTED_VALUE_INPUT).sendKeys("2022-04-01", Keys.ENTER);
 
+        // Constraint validation fails:
         testField.setInputValue("2/1/2022");
         assertClientInvalid();
         assertServerInvalid();
+        assertErrorMessage("");
 
+        // Binder validation fails:
         testField.setInputValue("3/1/2022");
-        assertClientValid();
-        assertServerValid();
+        assertClientInvalid();
+        assertServerInvalid();
+        assertErrorMessage(UNEXPECTED_VALUE_ERROR_MESSAGE);
 
+        // Both validations pass:
         testField.setInputValue("4/1/2022");
         assertClientValid();
         assertServerValid();
@@ -83,15 +71,21 @@ public class DatePickerValidationBasicIT
     @Test
     public void max_changeInputValue_assertValidity() {
         $("input").id(MAX_INPUT).sendKeys("2022-03-01", Keys.ENTER);
+        $("input").id(EXPECTED_VALUE_INPUT).sendKeys("2022-02-01", Keys.ENTER);
 
+        // Constraint validation fails:
         testField.setInputValue("4/1/2022");
         assertClientInvalid();
         assertServerInvalid();
+        assertErrorMessage("");
 
+        // Binder validation fails:
         testField.setInputValue("3/1/2022");
-        assertClientValid();
-        assertServerValid();
+        assertClientInvalid();
+        assertServerInvalid();
+        assertErrorMessage(UNEXPECTED_VALUE_ERROR_MESSAGE);
 
+        // Both validations pass:
         testField.setInputValue("2/1/2022");
         assertClientValid();
         assertServerValid();
@@ -99,9 +93,12 @@ public class DatePickerValidationBasicIT
 
     @Test
     public void badInput_changeInputValue_assertValidity() {
+        $("input").id(EXPECTED_VALUE_INPUT).sendKeys("2022-01-01", Keys.ENTER);
+
         testField.setInputValue("INVALID");
         assertServerInvalid();
         assertClientInvalid();
+        assertErrorMessage("");
 
         testField.setInputValue("1/1/2022");
         assertServerValid();
@@ -110,6 +107,7 @@ public class DatePickerValidationBasicIT
         testField.setInputValue("INVALID");
         assertServerInvalid();
         assertClientInvalid();
+        assertErrorMessage("");
     }
 
     protected DatePickerElement getTestField() {
